Upgrade JSesh dependency and only use library artifact
it is important to specify the specific artifact in the jitpack URL
because jitpack will otherwise generate an empty artifact for the jsesh
root pom and name it `jsesh` too. This results in 2 jars, one for the
jsesh library artifact and the empty one for the entire jsesh project,
but both named exactly the same. Both jars will find their way into
`.jar` and `.war` packages, where the classloader might pick the wrong
one. Interestingly, picking the wrong jar seems to happen only inside a
docker container...