store language selection in session
when users switch the UI translation language using the `?lang`
parameter, the selection now gets stored for the duration of their
browsing session, without the need to specify `?lang` in subsequent
requests.

This should fix some of the issues with inconsistent UI localization
currently experienced, as attested in e.g.
https://github.com/thesaurus-linguae-aegyptiae/tla-web/issues/37