add redundant routes to search handlers
make object type search results available under the additional path
`/search/{objectType}`. The existing routes of the form
`/{objectType}/search` are ambiguous because of the corresponding object details page
routes (`/{objectType}/{id}`).

The approach implemented with the procedural handler mapping might need
to be re-evaluated, because it seems to make tests act weird..